requirements.txt
streamlit==1.17.0
tensorflow-cpu==2.7.0
numpy==1.23.5
mecab-python3==1.0.3